+++ src/scenes/ray_tracing/RayTracer.cpp
@@ -250,7 +250,7 @@ void rayTrace(const std::vector<Object> &objects, const Light &light, const glm:
         }
 
 #pragma omp parallel for
-        for (size_t i = 0; i < rays.size(); i++) {
+        for (int i = 0; i < (int)rays.size(); i++) {
             pixels[i] = trace(rays[i], light, cameraPosition, objects, 0, maxRayDepth);
         }
